requirements.txt
Django==3.1.1
djangorestframework==3.13.1
scikit_learn==1.1.1
